% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_database}
\alias{connect_database}
\title{Database connection}
\usage{
connect_database(dbname, user, password, host, port = 15432)
}
\arguments{
\item{dbname}{name of database}

\item{user}{name of user}

\item{password}{password of database}

\item{host}{name of host}

\item{port}{number of port; default: 15432}
}
\value{
connection to database
}
\description{
Function \code{connect_database} establishes a connection with a PostgreSQL database.
}
\examples{
\dontrun{
connect_database(dbname, user, password, host, 15432)
}

}
\author{
Kacper Roszczyna, Dawid Stlemach
}

