% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bets_history.R
\name{create_bets_history}
\alias{create_bets_history}
\title{Table with bets}
\usage{
create_bets_history(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\description{
Function \code{create_bets_history} creates a table containing bets in a PostgreSQL database.
}
\examples{
\dontrun{
create_bets_history(connection)}

}
\author{
Kacper Roszczyna, Dawid Stelmach, Ania Elsner
}

