% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_database.R
\name{create_database}
\alias{create_database}
\title{Create all database tables}
\usage{
create_database(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\description{
Function \code{create_database} creates database containing tables: teams_names,
seasons, teams_seasons_statistics, games_story, bets_history.
}
\examples{
\dontrun{
create_database(connection)
}

}
\author{
Kacper Roszczyna Dawid Stelmach Piotr Smuda
}

