% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_games_story.R
\name{create_games_story}
\alias{create_games_story}
\title{Table with games history}
\usage{
create_games_story(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\description{
Function \code{create_games_story} creates a table containing games history in a PostgreSQL database.
}
\examples{
\dontrun{
create_games_story(connection)
}

}
\author{
Kacper Roszczyna Dawid Stelmach
}

