% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seasons.R
\name{create_seasons}
\alias{create_seasons}
\title{Table with seasons}
\usage{
create_seasons(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\description{
Function \code{create_seasons} creates a table containing seasons in a PostgreSQL database.
}
\examples{
\dontrun{
create_seasons(connection)
}

}
\author{
Kacper Roszczyna Dawid Stelmach
}

