% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_teams_names.R
\name{create_teams_names}
\alias{create_teams_names}
\title{Table with teams}
\usage{
create_teams_names(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\description{
Function \code{create_teams_names} creates a table containing teams in a PostgreSQL database.
}
\examples{
\dontrun{
create_teams_names(connection)
}

}
\author{
Kacper Roszczyna Dawid Stelmach
}

