% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_teams_seasons_statistics.R
\name{create_teams_seasons_statistics}
\alias{create_teams_seasons_statistics}
\title{Table with team season statistics}
\usage{
create_teams_seasons_statistics(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\description{
Function \code{create_teams_seasons_statistics} creates a table containing
teams' seasons statistics in a PostgreSQL database.
}
\examples{
\dontrun{
create_teams_seasons_statistics(connection)}

}
\author{
Kacper Roszczyna Dawid Stelmach
}

