% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect.R
\name{disconnect_database}
\alias{disconnect_database}
\title{Disconnect database}
\usage{
disconnect_database(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\description{
Function \code{disconnect_database} removes connection with a PostgreSQL database.
}
\examples{
\dontrun{
disconnect_database(connection)
}

}
\author{
Kacper Roszczyna
}

