% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_database.R
\name{drop_database}
\alias{drop_database}
\title{Dropping tables}
\usage{
drop_database(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\description{
Function \code{drop_database} removes all tables from PostgreSQL database.
}
\examples{
\dontrun{
drop_database(connection)
}

}
\author{
Kacper Roszczyna Piotr Smuda
}

