% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bets_history.R
\name{get_bets_history}
\alias{get_bets_history}
\title{Data frame with bookmaker statistics for NBA teams}
\usage{
get_bets_history(path)
}
\arguments{
\item{path}{path where phantomjs.exe file is contained}
}
\description{
Function \code{get_bets_history} downloads nba games bets from  http://www.oddsportal.com website
and creates data frame with them for all NBA teams.
Data frame bookmaker statistics from 2011 to 2016.
}
\note{
PhantomJS is needed to proper operation of this function.
You can download it from http://phantomjs.org/download.html.
}
\examples{
\dontrun{
get_bets_history(path)
}

}
\author{
Ania Elsner, Piotr Smuda
}

