% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games_story.R
\name{get_games_story}
\alias{get_games_story}
\title{Data frame with games history}
\usage{
get_games_story()
}
\description{
Function \code{get_games_story} downloads game statistics from the basketball-reference.com website
and creates data frame with them for all NBA teams.
Data frame contains games history from 2011 to 2016.
}
\examples{
\dontrun{
get_games_story()
}

}
\author{
Szymon Bazyluk, Aliaksandr Panimash, Piotr Smuda
}

