% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams_names.R
\name{get_teams_names}
\alias{get_teams_names}
\title{Data frame with team names}
\usage{
get_teams_names()
}
\description{
Function \code{get_teams_names} creates data frame with all NBA team names and their abbreviation.
}
\examples{
\dontrun{
get_teams_names()
}

}
\author{
Szymon Bazyluk, Aliaksandr Panimash
}

