% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams_seasons_statistics.R
\name{get_teams_seasons_statistics}
\alias{get_teams_seasons_statistics}
\title{Data frame with teams' statistics}
\usage{
get_teams_seasons_statistics()
}
\description{
Function \code{get_teams_seasons_statistics} downloads team statistics from
the basketball-reference.com website and creates data frame with them for all
NBA teams.
Data frame contains statistics from 2011 to 2016.
}
\examples{
\dontrun{
get_teams_seasons_statistics()
}

}
\author{
Szymon Bazyluk, Aliaksandr Panimash, Piotr Smuda
}

