% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bets_history_table.R
\name{read_bets_history_table}
\alias{read_bets_history_table}
\title{Read bets_history table from database}
\usage{
read_bets_history_table(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\value{
data.frame
}
\description{
Function \code{read_bets_history_table} read bets_history data frame with all
all best of NBA teams games from seasons in seasons table.
}
\examples{
\dontrun{
read_bets_history_table(connection)
}

}
\author{
Piotr Smuda
}

