% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_database.R
\name{read_database}
\alias{read_database}
\title{Read all database tables}
\usage{
read_database(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\description{
Function \code{read_database} reads all database tables: teams_names,
seasons, teams_seasons_statistics, games_story, bets_history.
}
\examples{
\dontrun{
read_database(connection)
}

}
\author{
Piotr Smuda
}

