% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_games_story_table.R
\name{read_games_story_table}
\alias{read_games_story_table}
\title{Read games_story table from database}
\usage{
read_games_story_table(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\value{
data.frame
}
\description{
Function \code{read_games_story_table} read games_story data frame with all
story of NBA teams games from seasons in seasons table.
}
\examples{
\dontrun{
read_games_story_table(connection)
}

}
\author{
Piotr Smuda
}

