% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seasons_table.R
\name{read_seasons_table}
\alias{read_seasons_table}
\title{Read seasons table from database}
\usage{
read_seasons_table(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\value{
data.frame
}
\description{
Function \code{read_seasons_table} read seasons data frame with seasons years.
}
\examples{
\dontrun{
read_seasons_table(connection)
}

}
\author{
Piotr Smuda
}

