% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_teams_names_table.R
\name{read_teams_names_table}
\alias{read_teams_names_table}
\title{Read teams_names table from database}
\usage{
read_teams_names_table(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\value{
data.frame
}
\description{
Function \code{read_teams_names_table} read teams_names data frame with all
NBA team names and their abbreviation.
}
\examples{
\dontrun{
read_teams_names_table(connection)
}

}
\author{
Piotr Smuda
}

