% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_teams_seasons_statistics_table.R
\name{read_teams_seasons_statistics_table}
\alias{read_teams_seasons_statistics_table}
\title{Read teams_seasons_statistics table from database}
\usage{
read_teams_seasons_statistics_table(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\value{
data.frame
}
\description{
Function \code{read_teams_seasons_statistics_table} read teams_seasons_statistics
data frame with all NBA teams seasons statistics from seasons in seasons table.
}
\examples{
\dontrun{
read_teams_seasons_statistics_table(connection)
}

}
\author{
Piotr Smuda
}

