% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_environment.R
\name{setup_environment}
\alias{setup_environment}
\title{Set up the environment}
\usage{
setup_environment()
}
\description{
Function \code{setup_environment} sets up whole work environment on PostgreSQL database.
}
\examples{
\dontrun{
setup_environment()
}

}
\author{
Kacper Roszczyna
}

