% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bets_history.R
\name{write_bets_history}
\alias{write_bets_history}
\title{Adding historical odds to database}
\usage{
write_bets_history(connection, clean_up = FALSE)
}
\arguments{
\item{connection}{connection to the database}

\item{path}{where phantomjs.exe file is contained}

\item{clean_up}{parameter if the data frame will be removed after adding to the database; default: FALSE}
}
\description{
Function \code{write_bets_history} adds historical odds to the PostgreSQL database.
}
\examples{
\dontrun{
write_bets_history(connection, getwd(), TRUE)
write_bets_history(connection, getwd(), FALSE)
}

}
\author{
Aliaksandr Panimash, Dawid Stelmach, Ania Elsner
}

