% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_database.R
\name{write_database}
\alias{write_database}
\title{Write all database tables}
\usage{
write_database(connection, path = getwd(), clean_up_teams_names = FALSE,
clean_up_teams_names = FALSE, clean_up_games_story = FALSE,
clean_up_team_season_stats = FALSE, clean_up_bets_history = FALSE)
}
\arguments{
\item{connection}{connection to the database}

\item{path}{where phantomjs.exe file is contained}

\item{clean_up_teams_names}{parameter if the teams_names data frame will
be removed after adding to the database; default: FALSE}

\item{clean_up_games_story}{parameter if the games_story data frame will
be removed after adding to the database; default: FALSE}

\item{clean_up_team_season_stats}{parameter if the team_season_stats data
frame will be removed after adding to the database; default: FALSE}

\item{clean_up_bets_history}{parameter if the bets_history data frame will
be removed after adding to the database; default: FALSE}
}
\description{
Function \code{write_database} writes database containing tables: teams_names,
seasons, teams_seasons_statistics, games_story, bets_history.
}
\examples{
\dontrun{
write_database(connection)
}

}
\author{
Piotr Smuda
}

