% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_games_story.R
\name{write_games_story}
\alias{write_games_story}
\title{Adding games story to database}
\usage{
write_games_story(connection, clean_up = FALSE)
}
\arguments{
\item{connection}{connection to the database}

\item{clean_up}{parameter if the data frame will be removed after adding to the database; default: FALSE}
}
\description{
Function \code{write_games_story} adds games story to the PostgreSQL database.
}
\examples{
\dontrun{
write_games_story(connection, TRUE)
write_games_story(connection, FALSE)
}

}
\author{
Aliaksandr Panimash, Dawid Stelmach
}

