% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_seasons.R
\name{write_seasons}
\alias{write_seasons}
\title{Adding seasons to database}
\usage{
write_seasons(connection)
}
\arguments{
\item{connection}{connection to the database}
}
\description{
Function \code{write_seasons} adds seasons to the PostgreSQL database.
}
\examples{
\dontrun{
write_seasons(connection)
}

}
\author{
Aliaksandr Panimash, Dawid Stelmach
}

