% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_teams_names.R
\name{write_teams_names}
\alias{write_teams_names}
\title{Adding team names to database}
\usage{
write_teams_names(connection, clean_up = FALSE)
}
\arguments{
\item{connection}{connection to the database}

\item{clean_up}{parameter if the data frame will be removed after adding to the database; default: FALSE}
}
\description{
Function \code{write_teams_names} adds team names to the PostgreSQL database.
}
\examples{
\dontrun{
write_teams_names(connection, TRUE)
write_teams_names(connection, FALSE)
}

}
\author{
Aliaksandr Panimash, Dawid Stelmach
}

