% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_teams_seasons_statistics.R
\name{write_teams_seasons_statistics}
\alias{write_teams_seasons_statistics}
\title{Adding team statistics to database}
\usage{
write_teams_seasons_statistics(connection, clean_up = FALSE)
}
\arguments{
\item{connection}{connection to the database}

\item{clean_up}{parameter if the data frame will be removed after adding to the database; default: FALSE}
}
\description{
Function \code{write_teams_seasons_statistics} adds team season statistics to the PostgreSQL database.
}
\examples{
\dontrun{
write_teams_seasons_statistics(connection, TRUE)
write_teams_seasons_statistics(connection, FALSE)
}

}
\author{
Aliaksandr Panimash, Dawid Stelmach
}

